+++ starlight/src/awi_structs/lazy_awi.rs
@@ -14,7 +14,7 @@ use crate::{
     awi,
     ensemble::{BasicValue, BasicValueKind, CommonValue, Ensemble, PExternal},
     epoch::get_current_epoch,
-    misc::DisplayStr,
+    utils::DisplayStr,
     Delay, Error, EvalAwi,
 };