+++ starlight/src/ensemble/optimize.rs
@@ -11,8 +11,8 @@ use awint::{
 
 use crate::{
     ensemble::{DynamicValue, Ensemble, LNode, LNodeKind, PBack, PLNode, PTNode, Referent, Value},
-    misc::SmallMap,
     triple_arena::{ptr_struct, OrdArena},
+    utils::SmallMap,
     Error,
 };