+++ starlight/src/utils.rs
@@ -1,10 +1,12 @@
 mod error;
 mod grid;
+mod ortho;
 mod rng;
 mod small_map;
 
 pub(crate) use error::DisplayStr;
 pub use error::Error;
-pub use grid::{Direction, Grid};
+pub use grid::Grid;
+pub use ortho::{Ortho, OrthoArray};
 pub use rng::StarRng;
 pub use small_map::{binary_search_similar_by, SmallMap, SmallSet};