+++ testcrate/tests/routing/pure.rs
@@ -1,0 +1,4 @@
+// pure routing with no combinatorics
+
+// TODO in another file test routing an example state machine over an island fabric with NAND static LUTs only, garbage, routing through inversion etc.
+