+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 - Removed the `Clone` impl for `EvalAwi` because it had questionable semantics and bad implications
 - `Epoch::suspend` now panics like it did in practice and does not return a `Result`
 - `EvalAwi`s and `LazyAwi`s no longer have any conditions under which they should panic on drop
+- Regrouped some things into a `utils` module
 
 ### Additions
 - `Loop`s and `Net`s can now have an arbitrary initial value
+- Added `Grid`, `Ortho`, and `OrthoArray`
 
 ## [0.3.0] - 2024-01-22
 ### Fixes